+++ src/redux/slices/alarm/alarmsAction.ts
@@ -1,0 +1,5 @@
+import generateRequest from '../../generalActions';
+
+import { type Alarms } from './alarmsSlice';
+
+export const getAlarms = generateRequest<Alarms>('get', 'alarms');